refactor: RSS3
