+++ packages/web3-providers/src/entry.ts
@@ -11,7 +11,7 @@ import {
 import { ZoraAPI } from './Zora/index.js'
 import { EtherscanExplorerAPI } from './Etherscan/index.js'
 import { RiskWarningAPI } from './RiskWarning/index.js'
-import { RSS3API } from './RSS3/index.js'
+import { RSS3API, RSS3NonFungibleTokenAPI } from './RSS3/index.js'
 import { TwitterAPI } from './Twitter/index.js'
 import { R2D2KeyValueAPI, R2D2DomainAPI, R2D2TokenListAPI } from './R2D2/index.js'
 import { InstagramAPI } from './Instagram/index.js'
@@ -70,7 +70,6 @@ export const CoinGeckoPriceEVM = new CoinGeckoPriceEVM_API()
 export const CoinGeckoPriceSolana = new CoinGeckoPriceSolanaAPI()
 export const EtherscanExplorer = new EtherscanExplorerAPI()
 export const RiskWarning = new RiskWarningAPI()
-export const RSS3 = new RSS3API()
 export const Twitter = new TwitterAPI()
 export const Mask = new MaskAPI()
 export const MaskX = new MaskX_API()
@@ -88,6 +87,10 @@ export const R2D2KeyValue = new R2D2KeyValueAPI()
 export const ENS = new ENS_API()
 export const SpaceID = new SpaceID_API()
 export const Bonfida = new BonfidaAPI()
+
+// RSS3
+export const RSS3 = new RSS3API()
+export const RSS3NonFungibleToken = new RSS3NonFungibleTokenAPI()
 
 // Debank
 export const DeBankGasOption = new DeBankGasOptionAPI()
